% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.GroupMerge.R
\name{checkGroupMerge}
\alias{checkGroupMerge}
\title{ICL Sweep}
\usage{
checkGroupMerge(Z, Y, g1, g2, alpha_var, beta_var)
}
\arguments{
\item{ICL}{value, Z, Y, G, alpha and beta}
}
\value{
Z_max, ICL_max
}
\description{
ICL Sweep
}
